+++ distribution_diagnostics.py
@@ -124,7 +124,7 @@ def main():
         for b in range(0, len(val_idx), BATCH):
             idx = val_idx[b:b + BATCH]
             x_num = ds.X_num[idx].to(DEVICE)
-            x_img = ds.X_img[idx].to(DEVICE)
+            x_img = ds.X_img[idx].to(DEVICE).float()
             x_txt = ds.X_txt[idx].to(DEVICE)
             y = ds.y[idx]
